backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
qdrant-client==1.12.0
psycopg2-binary==2.9.11
sqlalchemy==2.0.36
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.20
python-dotenv==1.0.1
httpx==0.28.1
pyjwt==2.9.0
cryptography==43.0.0
openai==1.55.0
cohere==5.11.0
google-generativeai==0.8.3
bcrypt==4.2.0
slowapi==0.1.9
tenacity==8.2.3